requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyJwt